JIRAT-87 [kotlin] charts

+++ src/main/java/br/com/jiratorio/domain/chart/IssueCountBySize.kt
@@ -2,5 +2,5 @@ package br.com.jiratorio.domain.chart
 
 data class IssueCountBySize(
     val labels: Set<String>,
-    val datasources: Map<String, List<Long>>
+    val datasources: Map<String, List<Int>>
 )
