+++ src/main/java/br/com/jiratorio/domain/response/issueperiod/IssuePeriodDetailResponse.kt
@@ -8,25 +8,17 @@ data class IssuePeriodDetailResponse(
     val dates: String,
     val leadTime: Double,
     val issuesCount: Int,
-
     val leadTimeByEstimate: Chart<String, Double>?,
-    val throughputByEstimate: Chart<String, Long>?,
-
+    val throughputByEstimate: Chart<String, Int>?,
     val leadTimeBySystem: Chart<String, Double>?,
-    val throughputBySystem: Chart<String, Long>?,
-
+    val throughputBySystem: Chart<String, Int>?,
     val leadTimeByType: Chart<String, Double>?,
-    val throughputByType: Chart<String, Long>?,
-
+    val throughputByType: Chart<String, Int>?,
     val leadTimeByProject: Chart<String, Double>?,
-    val throughputByProject: Chart<String, Long>?,
-
+    val throughputByProject: Chart<String, Int>?,
     val leadTimeByPriority: Chart<String, Double>?,
-    val throughputByPriority: Chart<String, Long>?,
-
+    val throughputByPriority: Chart<String, Int>?,
     val columnTimeAvg: MutableList<ColumnTimeAvg>?,
-
     val leadTimeCompareChart: Chart<String, Double>?,
-
     val dynamicCharts: MutableList<DynamicChart>?
 )
